:bug: Signer.sign() is an async method so the typing should indicate it can be a promise
At the moment the typings are wrong because the sign function is reported to be `sign(pdfBuffer: Buffer) => Buffer`.
However, this is not a correct typing as the method is expected to return a promise as it is `async`. The sign method
can also return a Buffer (sync) as is done in the P12Signer. Therefore the typings should be
`sign(pdfBuffer: Buffer) => Promise<Buffer> | Buffer`.